naniyamni_api/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cloudinary==1.44.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-polymorphic==4.1.0
djangorestframework==3.16.1
drf-polymorphic==2.1.1
psycopg2-binary==2.9.10
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
